pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "flask>=3.1.0",
    "psycopg2-binary>=2.9.10",
    "pypdf>=5.2.0",
]
